Airbyte/ingestion/requirements.txt
# HTTP client for Airbyte Cloud API calls
httpx

# Load .env variables
python-dotenv
